Add files via upload
The shell file "start_gradio.sh" executes the python file "radio_script.py". This script provides the UI for the AnimatedDrawings repository. 
It helps the user to generate the yaml files and also directly exports it as video or gif.

The user just has to drop his files into the default folders:
1. The character folder into .../AnimatedDrawings/examples/characters
2. The motion .yaml file into .../AnimatedDrawings/config/motion
3. The target .yaml file into .../AnimatedDrawings/config/retarget